scripts: west_commands: runners: fix west rtt for shell usage
When west rtt command is executed two things happen:
- Some variation of GDB server is run.
- A telnet client (via sockets) is opened to connect to that server.

Previously, that telnet client was operating in the line mode in both
directions. That mode is perfectly fine for just presenting the captured
content (RTT data) to the stdout, however it is not suitable for
interacting with the shell on the device.

None of the control commands, such as tab, arrow keys, are correctly
passed down to the device.

By switching the telnet client into the character mode now this works as
expected.